Read current function addresses in the IAT instead of original
Previously the `resolved_address` of an imported function was taken from
the address given by the matching export. Sometimes, however, the IAT
will be modified by another DLL loading (e.g. wsock32.dll which
redirects all its calls to ws2_32.dll). This means that the
`resolved_address` can point to an old function (or in the case of
wsock32.dll, to a string instead of a function.

Now we instead take the `resolved_address` to be the value at `thunk`,
and use `original_address` for what we would have resolved the import
to be be.